+++ CHANGELOG.md
@@ -25,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [3.6.3] - 2024.08.02 - 18496f1bdf4faa0afa7f2e4a824c851e40b57f0f
 ### Added
 - Update of Offer claims
-- Checks for issuer & data data in a did:peer JWANT
+- Checks for issuer & data in a did:peer JWANT
 ### Changed
 - Consolidate DID crypto facilities into "vc" directory.
 ### Removed